app/requirements.txt
aiokafka==0.12.0
environs==10.3.0
python-dotenv==1.0.1